refactor: extract common database logic into RelationalErrorLog base class
MAINT-002: Reduce code duplication across database implementations.

- Add RelationalErrorLog abstract base class in ElmahCore
  - Common implementations: Log, GetError, GetErrors
  - Common async implementations: LogAsync, GetErrorAsync, GetErrorsAsync
  - Abstract methods for database-specific operations:
    CreateConnection, CreateLogErrorCommand, CreateGetErrorXmlCommand,
    CreateGetErrorsXmlCommand, CreateGetErrorsCountCommand, CreateTableIfNotExists

- Refactor SqlErrorLog to inherit from RelationalErrorLog
- Refactor MySqlErrorLog to inherit from RelationalErrorLog
- Refactor PgsqlErrorLog to inherit from RelationalErrorLog
- Delete unused CommandExtension.cs from MySql project

Reduces total code by ~600 lines while maintaining all functionality.
