+++ ElmahCore.MsSql/SqlErrorLog.cs
@@ -1,8 +1,6 @@
 using System;
-using System.Collections.Generic;
 using System.Data;
-using System.Threading;
-using System.Threading.Tasks;
+using System.Data.Common;
 using Microsoft.Data.SqlClient;
 using Microsoft.Extensions.Options;
 
@@ -13,10 +11,13 @@ namespace ElmahCore.Sql;
 ///     as its backing store.
 /// </summary>
 // ReSharper disable once UnusedType.Global
-public class SqlErrorLog : ErrorLog
+public class SqlErrorLog : RelationalErrorLog
 {
     private const int MaxAppNameLength = 60;
     private readonly bool _logAllXml;
+    private readonly string _connectionString;
+    private readonly string _schemaName;
+    private readonly string _tableName;
 
     /// <summary>
     ///     Initializes a new instance of the <see cref="SqlErrorLog" /> class
@@ -39,272 +40,145 @@ public class SqlErrorLog : ErrorLog
         if (string.IsNullOrEmpty(connectionString))
             throw new ArgumentNullException(nameof(connectionString));
 
-        ConnectionString = connectionString;
-        DatabaseSchemaName = !string.IsNullOrWhiteSpace(schemaName) ? schemaName : "dbo";
-        DatabaseTableName = !string.IsNullOrWhiteSpace(tableName) ? tableName : "ELMAH_Error";
+        _connectionString = connectionString;
+        _schemaName = !string.IsNullOrWhiteSpace(schemaName) ? schemaName : "dbo";
+        _tableName = !string.IsNullOrWhiteSpace(tableName) ? tableName : "ELMAH_Error";
         _logAllXml = logAllXml;
 
         if (createTablesIfNotExist)
             CreateTableIfNotExists();
     }
 
-    /// <summary>
-    ///     Gets the name of this error log implementation.
-    /// </summary>
+    /// <inheritdoc />
     public override string Name => "MSSQL Error Log";
 
-    /// <summary>
-    ///     Gets the connection string used by the log to connect to the database.
-    /// </summary>
-    // ReSharper disable once MemberCanBeProtected.Global
-    public virtual string ConnectionString { get; }
+    /// <inheritdoc />
+    protected override string ConnectionString => _connectionString;
+
+    /// <inheritdoc />
+    protected override bool LogAllXml => _logAllXml;
 
     /// <summary>
-    /// Gets the Schema name to be used for the error table
+    ///     Gets the schema name to be used for the error table.
     /// </summary>
-    protected virtual string DatabaseSchemaName { get; }
+    protected virtual string DatabaseSchemaName => _schemaName;
 
     /// <summary>
-    /// Gets the Table name to be used for the error table
+    ///     Gets the table name to be used for the error table.
     /// </summary>
-    protected virtual string DatabaseTableName { get; }
+    protected virtual string DatabaseTableName => _tableName;
 
-    public override string Log(Error error)
+    /// <inheritdoc />
+    protected override DbConnection CreateConnection() => new SqlConnection(ConnectionString);
+
+    /// <inheritdoc />
+    protected override DbCommand CreateLogErrorCommand(Guid id, string appName, string hostName, string typeName,
+        string source, string message, string user, int statusCode, DateTime time, string xml)
     {
-        var id = Guid.NewGuid();
-        Log(id, error);
-        return id.ToString();
+        var command = new SqlCommand
+        {
+            CommandText = $@"
+/* elmah */
+INSERT INTO [{DatabaseSchemaName}].[{DatabaseTableName}] (ErrorId, Application, Host, Type, Source, Message, ""User"", StatusCode, TimeUtc, AllXml)
+VALUES (@ErrorId, @Application, @Host, @Type, @Source, @Message, @User, @StatusCode, @TimeUtc, @AllXml)
+"
+        };
+        command.Parameters.Add(new SqlParameter("ErrorId", id));
+        command.Parameters.Add(new SqlParameter("Application", appName));
+        command.Parameters.Add(new SqlParameter("Host", hostName));
+        command.Parameters.Add(new SqlParameter("Type", typeName));
+        command.Parameters.Add(new SqlParameter("Source", source));
+        command.Parameters.Add(new SqlParameter("Message", message));
+        command.Parameters.Add(new SqlParameter("User", user));
+        command.Parameters.Add(new SqlParameter("StatusCode", statusCode));
+        command.Parameters.Add(new SqlParameter("TimeUtc", time.ToUniversalTime()));
+        command.Parameters.Add(new SqlParameter("AllXml", xml));
+
+        return command;
     }
 
-    public override void Log(Guid id, Error error)
+    /// <inheritdoc />
+    protected override DbCommand CreateGetErrorXmlCommand(string appName, Guid errorId)
     {
-        try
+        var command = new SqlCommand
         {
-            var errorXml = _logAllXml
-                ? ErrorXml.EncodeString(error)
-                : "<error message=\"AllXml logging disabled\" />";
+            CommandText = $@"
+SELECT AllXml FROM [{DatabaseSchemaName}].[{DatabaseTableName}]
+WHERE
+    Application = @Application
+    AND ErrorId = @ErrorId
+"
+        };
 
-            using var connection = new SqlConnection(ConnectionString);
-            using var command = Commands.LogError(id, ApplicationName, error.HostName, error.Type, error.Source,
-                error.Message, error.User, error.StatusCode, error.Time, errorXml,
-                DatabaseSchemaName, DatabaseTableName);
-            command.Connection = connection;
-            connection.Open();
-            command.ExecuteNonQuery();
-        }
-        catch
-        {
-            //guard: silently fail, this can't bubble up or it will create a stack overflow from errors attempting to log errors....
-        }
+        command.Parameters.Add(new SqlParameter("Application", appName));
+        command.Parameters.Add(new SqlParameter("ErrorId", errorId));
+
+        return command;
     }
 
-    public override ErrorLogEntry GetError(string id)
+    /// <inheritdoc />
+    protected override DbCommand CreateGetErrorsXmlCommand(string appName, int errorIndex, int pageSize)
     {
-        ArgumentNullException.ThrowIfNull(id);
-        if (id.Length == 0) throw new ArgumentException(null, nameof(id));
-
-        Guid errorGuid;
-
-        try
+        var command = new SqlCommand
         {
-            errorGuid = new Guid(id);
-        }
-        catch (FormatException e)
-        {
-            throw new ArgumentException(e.Message, nameof(id), e);
-        }
+            CommandText = $@"
+SELECT ErrorId, AllXml FROM [{DatabaseSchemaName}].[{DatabaseTableName}]
+WHERE
+    Application = @Application
+ORDER BY [Sequence] DESC
+OFFSET     @offset ROWS
+FETCH NEXT @limit ROWS ONLY;
+"
+        };
 
-        string errorXml;
+        command.Parameters.Add("@Application", SqlDbType.NVarChar, MaxAppNameLength).Value = appName;
+        command.Parameters.Add("@offset", SqlDbType.Int).Value = errorIndex;
+        command.Parameters.Add("@limit", SqlDbType.Int).Value = pageSize;
 
-        using (var connection = new SqlConnection(ConnectionString))
-        using (var command = Commands.GetErrorXml(ApplicationName, errorGuid,
-                   DatabaseSchemaName, DatabaseTableName))
-        {
-            command.Connection = connection;
-            connection.Open();
-            errorXml = (string)command.ExecuteScalar();
-        }
-
-        if (errorXml == null)
-            return null;
-
-        var error = ErrorXml.DecodeString(errorXml);
-        return new ErrorLogEntry(this, id, error);
+        return command;
     }
 
-    public override int GetErrors(int errorIndex, int pageSize, ICollection<ErrorLogEntry> errorEntryList)
+    /// <inheritdoc />
+    protected override DbCommand CreateGetErrorsCountCommand(string appName)
     {
-        if (errorIndex < 0) throw new ArgumentOutOfRangeException(nameof(errorIndex), errorIndex, null);
-        if (pageSize < 0) throw new ArgumentOutOfRangeException(nameof(pageSize), pageSize, null);
+        var command = new SqlCommand
+        {
+            CommandText = $"SELECT COUNT(*) FROM [{DatabaseSchemaName}].[{DatabaseTableName}] WHERE Application = @Application"
+        };
+        command.Parameters.Add("@Application", SqlDbType.NVarChar, MaxAppNameLength).Value = appName;
+        return command;
+    }
 
+    /// <inheritdoc />
+    protected override void CreateTableIfNotExists()
+    {
         using var connection = new SqlConnection(ConnectionString);
         connection.Open();
 
-        using (var command = Commands.GetErrorsXml(ApplicationName, errorIndex, pageSize,
-                   DatabaseSchemaName, DatabaseTableName))
+        using var cmdCheck = new SqlCommand
         {
-            command.Connection = connection;
+            Connection = connection,
+            CommandText = $@"
+SELECT 1
+WHERE EXISTS (
+   SELECT 1
+   FROM   INFORMATION_SCHEMA.TABLES
+   WHERE  TABLE_SCHEMA = '{DatabaseSchemaName}'
+   AND    TABLE_NAME = '{DatabaseTableName}'
+   )
+"
+        };
 
-            using (var reader = command.ExecuteReader())
-            {
-                while (reader.Read())
-                {
-                    var id = reader.GetGuid(0);
-                    var xml = reader.GetString(1);
-                    var error = ErrorXml.DecodeString(xml);
-                    errorEntryList.Add(new ErrorLogEntry(this, id.ToString(), error));
-                }
-            }
-        }
-
-        using (var command = Commands.GetErrorsXmlTotal(ApplicationName,
-                   DatabaseSchemaName, DatabaseTableName))
-        {
-            command.Connection = connection;
-            return int.Parse(command.ExecuteScalar().ToString());
-        }
-    }
-
-    /// <inheritdoc />
-    public override async Task<string> LogAsync(Error error, CancellationToken cancellationToken)
-    {
-        var id = Guid.NewGuid();
-
-        try
-        {
-            var errorXml = _logAllXml
-                ? ErrorXml.EncodeString(error)
-                : "<error message=\"AllXml logging disabled\" />";
-
-            await using var connection = new SqlConnection(ConnectionString);
-            await using var command = Commands.LogError(id, ApplicationName, error.HostName, error.Type, error.Source,
-                error.Message, error.User, error.StatusCode, error.Time, errorXml,
-                DatabaseSchemaName, DatabaseTableName);
-            command.Connection = connection;
-            await connection.OpenAsync(cancellationToken);
-            await command.ExecuteNonQueryAsync(cancellationToken);
-        }
-        catch
-        {
-            //guard: silently fail, this can't bubble up or it will create a stack overflow from errors attempting to log errors....
-        }
-
-        return id.ToString();
-    }
-
-    /// <inheritdoc />
-    public override async Task<ErrorLogEntry> GetErrorAsync(string id, CancellationToken cancellationToken)
-    {
-        ArgumentNullException.ThrowIfNull(id);
-        if (id.Length == 0) throw new ArgumentException(null, nameof(id));
-
-        Guid errorGuid;
-
-        try
-        {
-            errorGuid = new Guid(id);
-        }
-        catch (FormatException e)
-        {
-            throw new ArgumentException(e.Message, nameof(id), e);
-        }
-
-        string errorXml;
-
-        await using (var connection = new SqlConnection(ConnectionString))
-        await using (var command = Commands.GetErrorXml(ApplicationName, errorGuid,
-                         DatabaseSchemaName, DatabaseTableName))
-        {
-            command.Connection = connection;
-            await connection.OpenAsync(cancellationToken);
-            errorXml = (string)await command.ExecuteScalarAsync(cancellationToken);
-        }
-
-        if (errorXml == null)
-            return null;
-
-        var error = ErrorXml.DecodeString(errorXml);
-        return new ErrorLogEntry(this, id, error);
-    }
-
-    /// <inheritdoc />
-    public override async Task<int> GetErrorsAsync(int errorIndex, int pageSize, ICollection<ErrorLogEntry> errorEntryList,
-        CancellationToken cancellationToken)
-    {
-        if (errorIndex < 0) throw new ArgumentOutOfRangeException(nameof(errorIndex), errorIndex, null);
-        if (pageSize < 0) throw new ArgumentOutOfRangeException(nameof(pageSize), pageSize, null);
-
-        await using var connection = new SqlConnection(ConnectionString);
-        await connection.OpenAsync(cancellationToken);
-
-        await using (var command = Commands.GetErrorsXml(ApplicationName, errorIndex, pageSize,
-                         DatabaseSchemaName, DatabaseTableName))
-        {
-            command.Connection = connection;
-
-            await using (var reader = await command.ExecuteReaderAsync(cancellationToken))
-            {
-                while (await reader.ReadAsync(cancellationToken))
-                {
-                    var id = reader.GetGuid(0);
-                    var xml = reader.GetString(1);
-                    var error = ErrorXml.DecodeString(xml);
-                    errorEntryList.Add(new ErrorLogEntry(this, id.ToString(), error));
-                }
-            }
-        }
-
-        await using (var command = Commands.GetErrorsXmlTotal(ApplicationName,
-                         DatabaseSchemaName, DatabaseTableName))
-        {
-            command.Connection = connection;
-            return int.Parse((await command.ExecuteScalarAsync(cancellationToken))?.ToString() ?? "0");
-        }
-    }
-
-    /// <summary>
-    ///  Creates the necessary tables and sequences used by this implementation
-    /// </summary>
-    private void CreateTableIfNotExists()
-    {
-        using var connection = new SqlConnection(ConnectionString);
-        connection.Open();
-        using var cmdCheck = Commands.CheckTable(DatabaseSchemaName, DatabaseTableName);
-        cmdCheck.Connection = connection;
-        // ReSharper disable once PossibleNullReferenceException
         var exists = (int?)cmdCheck.ExecuteScalar();
 
         if (!exists.HasValue)
-            ExecuteBatchNonQuery(Commands.CreateTableSql(DatabaseSchemaName, DatabaseTableName),
-                connection);
+            ExecuteBatchNonQuery(CreateTableSql(), connection);
     }
 
-    private static void ExecuteBatchNonQuery(string sql, SqlConnection conn)
+    private string CreateTableSql()
     {
-        var sqlBatch = string.Empty;
-        using var cmd = new SqlCommand(string.Empty, conn);
-        sql += "\nGO"; // make sure the last batch is executed.
-        foreach (var line in sql.Split(["\n", "\r"],
-                     StringSplitOptions.RemoveEmptyEntries))
-            if (line.ToUpperInvariant().Trim() == "GO")
-            {
-                cmd.CommandText = sqlBatch;
-                cmd.ExecuteNonQuery();
-                sqlBatch = string.Empty;
-            }
-            else
-            {
-                sqlBatch += line + "\n";
-            }
-    }
-
-    private static class Commands
-    {
-        public static string CreateTableSql(string schemaName, string tableName)
-        {
-            return
-                $@"
-CREATE TABLE [{schemaName}].[{tableName}]
+        return $@"
+CREATE TABLE [{DatabaseSchemaName}].[{DatabaseTableName}]
 (
     [ErrorId]     UNIQUEIDENTIFIER NOT NULL,
     [Application] NVARCHAR(60)  NOT NULL,
@@ -316,137 +190,44 @@ CREATE TABLE [{schemaName}].[{tableName}]
     [StatusCode]  INT NOT NULL,
     [TimeUtc]     DATETIME NOT NULL,
     [Sequence]    INT IDENTITY (1, 1) NOT NULL,
-    [AllXml]      NVARCHAR(MAX) NOT NULL 
-) 
+    [AllXml]      NVARCHAR(MAX) NOT NULL
+)
 GO
 
-ALTER TABLE [{schemaName}].[{tableName}] WITH NOCHECK ADD 
-    CONSTRAINT [PK_{tableName}] PRIMARY KEY NONCLUSTERED ([ErrorId]) ON [PRIMARY] 
+ALTER TABLE [{DatabaseSchemaName}].[{DatabaseTableName}] WITH NOCHECK ADD
+    CONSTRAINT [PK_{DatabaseTableName}] PRIMARY KEY NONCLUSTERED ([ErrorId]) ON [PRIMARY]
 GO
 
-ALTER TABLE [{schemaName}].[{tableName}] ADD 
-    CONSTRAINT [DF_{tableName}_ErrorId] DEFAULT (NEWID()) FOR [ErrorId]
+ALTER TABLE [{DatabaseSchemaName}].[{DatabaseTableName}] ADD
+    CONSTRAINT [DF_{DatabaseTableName}_ErrorId] DEFAULT (NEWID()) FOR [ErrorId]
 GO
 
-CREATE NONCLUSTERED INDEX [IX_{tableName}_App_Time_Seq] ON [{schemaName}].[{tableName}] 
+CREATE NONCLUSTERED INDEX [IX_{DatabaseTableName}_App_Time_Seq] ON [{DatabaseSchemaName}].[{DatabaseTableName}]
 (
     [Application]   ASC,
     [TimeUtc]       DESC,
     [Sequence]      DESC
-) 
+)
 ON [PRIMARY]";
-        }
+    }
 
-        public static SqlCommand CheckTable(string schemaName, string tableName)
+    private static void ExecuteBatchNonQuery(string sql, SqlConnection conn)
+    {
+        var sqlBatch = string.Empty;
+        using var cmd = new SqlCommand(string.Empty, conn);
+        sql += "\nGO"; // make sure the last batch is executed.
+        foreach (var line in sql.Split(["\n", "\r"], StringSplitOptions.RemoveEmptyEntries))
         {
-            return new SqlCommand
+            if (line.ToUpperInvariant().Trim() == "GO")
             {
-                CommandText = $@"
-SELECT 1 
-WHERE EXISTS (
-   SELECT 1
-   FROM   INFORMATION_SCHEMA.TABLES 
-   WHERE  TABLE_SCHEMA = '{schemaName}'
-   AND    TABLE_NAME = '{tableName}'
-   )
-"
-            };
-        }
-
-        public static SqlCommand LogError(
-            Guid id,
-            string appName,
-            string hostName,
-            string typeName,
-            string source,
-            string message,
-            string user,
-            int statusCode,
-            DateTime time,
-            string xml,
-            string schemaName,
-            string tableName)
-        {
-            var command = new SqlCommand
+                cmd.CommandText = sqlBatch;
+                cmd.ExecuteNonQuery();
+                sqlBatch = string.Empty;
+            }
+            else
             {
-                CommandText = $@"
-/* elmah */
-INSERT INTO [{schemaName}].[{tableName}] (ErrorId, Application, Host, Type, Source, Message, ""User"", StatusCode, TimeUtc, AllXml)
-VALUES (@ErrorId, @Application, @Host, @Type, @Source, @Message, @User, @StatusCode, @TimeUtc, @AllXml)
-"
-            };
-            command.Parameters.Add(new SqlParameter("ErrorId", id));
-            command.Parameters.Add(new SqlParameter("Application", appName));
-            command.Parameters.Add(new SqlParameter("Host", hostName));
-            command.Parameters.Add(new SqlParameter("Type", typeName));
-            command.Parameters.Add(new SqlParameter("Source", source));
-            command.Parameters.Add(new SqlParameter("Message", message));
-            command.Parameters.Add(new SqlParameter("User", user));
-            command.Parameters.Add(new SqlParameter("StatusCode", statusCode));
-            command.Parameters.Add(new SqlParameter("TimeUtc", time.ToUniversalTime()));
-            command.Parameters.Add(new SqlParameter("AllXml", xml));
-
-            return command;
-        }
-
-        public static SqlCommand GetErrorXml(
-            string appName,
-            Guid id,
-            string schemaName,
-            string tableName)
-        {
-            var command = new SqlCommand
-            {
-                CommandText = $@"
-SELECT AllXml FROM [{schemaName}].[{tableName}]
-WHERE 
-    Application = @Application 
-    AND ErrorId = @ErrorId
-"
-            };
-
-            command.Parameters.Add(new SqlParameter("Application", appName));
-            command.Parameters.Add(new SqlParameter("ErrorId", id));
-
-            return command;
-        }
-
-        public static SqlCommand GetErrorsXml(
-            string appName,
-            int errorIndex,
-            int pageSize,
-            string schemaName,
-            string tableName)
-        {
-            var command = new SqlCommand
-            {
-                CommandText = $@"
-SELECT ErrorId, AllXml FROM [{schemaName}].[{tableName}]
-WHERE
-    Application = @Application
-ORDER BY [Sequence] DESC
-OFFSET     @offset ROWS
-FETCH NEXT @limit ROWS ONLY;
-"
-            };
-
-            command.Parameters.Add("@Application", SqlDbType.NVarChar, MaxAppNameLength).Value = appName;
-            command.Parameters.Add("@offset", SqlDbType.Int).Value = errorIndex;
-            command.Parameters.Add("@limit", SqlDbType.Int).Value = pageSize;
-
-            return command;
-        }
-
-        public static SqlCommand GetErrorsXmlTotal(string appName,
-            string schemaName,
-            string tableName)
-        {
-            var command = new SqlCommand
-            {
-                CommandText = $"SELECT COUNT(*) FROM [{schemaName}].[{tableName}] WHERE Application = @Application"
-            };
-            command.Parameters.Add("@Application", SqlDbType.NVarChar, MaxAppNameLength).Value = appName;
-            return command;
+                sqlBatch += line + "\n";
+            }
         }
     }
 }
