+++ ElmahCore.Postgresql/PgsqlErrorLog.cs
@@ -1,7 +1,5 @@
 using System;
-using System.Collections.Generic;
-using System.Threading;
-using System.Threading.Tasks;
+using System.Data.Common;
 using JetBrains.Annotations;
 using Microsoft.Extensions.Options;
 using Npgsql;
@@ -14,17 +12,18 @@ namespace ElmahCore.Postgresql;
 ///     as its backing store.
 /// </summary>
 [UsedImplicitly]
-public class PgsqlErrorLog : ErrorLog
+public class PgsqlErrorLog : RelationalErrorLog
 {
     private const int MaxAppNameLength = 60;
     private readonly bool _logAllXml;
+    private readonly string _connectionString;
 
     /// <summary>
     ///     Initializes a new instance of the <see cref="PgsqlErrorLog" /> class
     ///     using a dictionary of configured settings.
     /// </summary>
-    public PgsqlErrorLog(IOptions<ElmahOptions> option) : this(option.Value.ConnectionString, option.Value.CreateTablesIfNotExist,
-        option.Value.LogAllXml)
+    public PgsqlErrorLog(IOptions<ElmahOptions> option) : this(option.Value.ConnectionString,
+        option.Value.CreateTablesIfNotExist, option.Value.LogAllXml)
     {
     }
 
@@ -37,241 +36,132 @@ public class PgsqlErrorLog : ErrorLog
         if (string.IsNullOrEmpty(connectionString))
             throw new ArgumentNullException(nameof(connectionString));
 
-        ConnectionString = connectionString;
+        _connectionString = connectionString;
         _logAllXml = logAllXml;
 
         if (createTablesIfNotExist)
             CreateTableIfNotExists();
     }
 
-    /// <summary>
-    ///     Gets the name of this error log implementation.
-    /// </summary>
+    /// <inheritdoc />
     public override string Name => "PostgreSQL Error Log";
 
-    /// <summary>
-    ///     Gets the connection string used by the log to connect to the database.
-    /// </summary>
-    public virtual string ConnectionString { get; }
+    /// <inheritdoc />
+    protected override string ConnectionString => _connectionString;
 
-    public override string Log(Error error)
+    /// <inheritdoc />
+    protected override bool LogAllXml => _logAllXml;
+
+    /// <inheritdoc />
+    protected override DbConnection CreateConnection() => new NpgsqlConnection(ConnectionString);
+
+    /// <inheritdoc />
+    protected override DbCommand CreateLogErrorCommand(Guid id, string appName, string hostName, string typeName,
+        string source, string message, string user, int statusCode, DateTime time, string xml)
     {
-        var id = Guid.NewGuid();
-
-        Log(id, error);
-
-        return id.ToString();
-    }
-
-    public override void Log(Guid id, Error error)
-    {
-        ArgumentNullException.ThrowIfNull(error);
-
-        var errorXml = _logAllXml
-            ? ErrorXml.EncodeString(error)
-            : "<error message=\"AllXml logging disabled\" />";
-
-        using var connection = new NpgsqlConnection(ConnectionString);
-        using var command = Commands.LogError(id, ApplicationName, error.HostName, error.Type, error.Source,
-            error.Message, error.User, error.StatusCode, error.Time, errorXml);
-        command.Connection = connection;
-        connection.Open();
-        command.ExecuteNonQuery();
-    }
-
-    public override ErrorLogEntry GetError(string id)
-    {
-        ArgumentNullException.ThrowIfNull(id);
-        if (id.Length == 0) throw new ArgumentException(null, nameof(id));
-
-        Guid errorGuid;
-
-        try
+        var command = new NpgsqlCommand
         {
-            errorGuid = new Guid(id);
-        }
-        catch (FormatException e)
-        {
-            throw new ArgumentException(e.Message, nameof(id), e);
-        }
+            CommandText = @"
+/* elmah */
+INSERT INTO Elmah_Error (ErrorId, Application, Host, Type, Source, Message, ""User"", StatusCode, TimeUtc, AllXml)
+VALUES (@ErrorId, @Application, @Host, @Type, @Source, @Message, @User, @StatusCode, @TimeUtc, @AllXml)
+"
+        };
 
-        string errorXml;
+        command.Parameters.Add(new NpgsqlParameter("ErrorId", id));
+        command.Parameters.Add(new NpgsqlParameter("Application", appName));
+        command.Parameters.Add(new NpgsqlParameter("Host", hostName));
+        command.Parameters.Add(new NpgsqlParameter("Type", typeName));
+        command.Parameters.Add(new NpgsqlParameter("Source", source));
+        command.Parameters.Add(new NpgsqlParameter("Message", message));
+        command.Parameters.Add(new NpgsqlParameter("User", user));
+        command.Parameters.Add(new NpgsqlParameter("StatusCode", statusCode));
+        command.Parameters.Add(new NpgsqlParameter("TimeUtc", time.ToUniversalTime()));
+        command.Parameters.Add(new NpgsqlParameter("AllXml", xml));
 
-        using (var connection = new NpgsqlConnection(ConnectionString))
-        using (var command = Commands.GetErrorXml(ApplicationName, errorGuid))
-        {
-            command.Connection = connection;
-            connection.Open();
-            errorXml = (string)command.ExecuteScalar();
-        }
-
-        if (errorXml == null)
-            return null;
-
-        var error = ErrorXml.DecodeString(errorXml);
-        return new ErrorLogEntry(this, id, error);
-    }
-
-    public override int GetErrors(int errorIndex, int pageSize, ICollection<ErrorLogEntry> errorEntryList)
-    {
-        if (errorIndex < 0) throw new ArgumentOutOfRangeException(nameof(errorIndex), errorIndex, null);
-        if (pageSize < 0) throw new ArgumentOutOfRangeException(nameof(pageSize), pageSize, null);
-
-        using var connection = new NpgsqlConnection(ConnectionString);
-        connection.Open();
-
-        using (var command = Commands.GetErrorsXml(ApplicationName, errorIndex, pageSize))
-        {
-            command.Connection = connection;
-
-            using (var reader = command.ExecuteReader())
-            {
-                while (reader.Read())
-                {
-                    var id = reader.GetGuid(0);
-                    var xml = reader.GetString(1);
-                    var error = ErrorXml.DecodeString(xml);
-                    errorEntryList.Add(new ErrorLogEntry(this, id.ToString(), error));
-                }
-            }
-        }
-
-        using (var command = Commands.GetErrorsXmlTotal(ApplicationName))
-        {
-            command.Connection = connection;
-            return Convert.ToInt32(command.ExecuteScalar());
-        }
+        return command;
     }
 
     /// <inheritdoc />
-    public override async Task<string> LogAsync(Error error, CancellationToken cancellationToken)
+    protected override DbCommand CreateGetErrorXmlCommand(string appName, Guid errorId)
     {
-        ArgumentNullException.ThrowIfNull(error);
+        var command = new NpgsqlCommand
+        {
+            CommandText = @"
+SELECT AllXml
+FROM Elmah_Error
+WHERE
+    Application = @Application
+    AND ErrorId = @ErrorId
+"
+        };
 
-        var id = Guid.NewGuid();
+        command.Parameters.Add(new NpgsqlParameter("Application", appName));
+        command.Parameters.Add(new NpgsqlParameter("ErrorId", errorId));
 
-        var errorXml = _logAllXml
-            ? ErrorXml.EncodeString(error)
-            : "<error message=\"AllXml logging disabled\" />";
-
-        await using var connection = new NpgsqlConnection(ConnectionString);
-        await using var command = Commands.LogError(id, ApplicationName, error.HostName, error.Type, error.Source,
-            error.Message, error.User, error.StatusCode, error.Time, errorXml);
-        command.Connection = connection;
-        await connection.OpenAsync(cancellationToken);
-        await command.ExecuteNonQueryAsync(cancellationToken);
-
-        return id.ToString();
+        return command;
     }
 
     /// <inheritdoc />
-    public override async Task<ErrorLogEntry> GetErrorAsync(string id, CancellationToken cancellationToken)
+    protected override DbCommand CreateGetErrorsXmlCommand(string appName, int errorIndex, int pageSize)
     {
-        ArgumentNullException.ThrowIfNull(id);
-        if (id.Length == 0) throw new ArgumentException(null, nameof(id));
-
-        Guid errorGuid;
-
-        try
+        var command = new NpgsqlCommand
         {
-            errorGuid = new Guid(id);
-        }
-        catch (FormatException e)
-        {
-            throw new ArgumentException(e.Message, nameof(id), e);
-        }
+            CommandText = @"
+SELECT ErrorId, AllXml FROM Elmah_Error
+WHERE
+    Application = @Application
+ORDER BY Sequence DESC
+OFFSET @offset
+LIMIT @limit
+"
+        };
 
-        string errorXml;
+        command.Parameters.Add("@Application", NpgsqlDbType.Text, MaxAppNameLength).Value = appName;
+        command.Parameters.Add("@offset", NpgsqlDbType.Integer).Value = errorIndex;
+        command.Parameters.Add("@limit", NpgsqlDbType.Integer).Value = pageSize;
 
-        await using (var connection = new NpgsqlConnection(ConnectionString))
-        await using (var command = Commands.GetErrorXml(ApplicationName, errorGuid))
-        {
-            command.Connection = connection;
-            await connection.OpenAsync(cancellationToken);
-            errorXml = (string)await command.ExecuteScalarAsync(cancellationToken);
-        }
-
-        if (errorXml == null)
-            return null;
-
-        var error = ErrorXml.DecodeString(errorXml);
-        return new ErrorLogEntry(this, id, error);
+        return command;
     }
 
     /// <inheritdoc />
-    public override async Task<int> GetErrorsAsync(int errorIndex, int pageSize, ICollection<ErrorLogEntry> errorEntryList,
-        CancellationToken cancellationToken)
+    protected override DbCommand CreateGetErrorsCountCommand(string appName)
     {
-        if (errorIndex < 0) throw new ArgumentOutOfRangeException(nameof(errorIndex), errorIndex, null);
-        if (pageSize < 0) throw new ArgumentOutOfRangeException(nameof(pageSize), pageSize, null);
-
-        await using var connection = new NpgsqlConnection(ConnectionString);
-        await connection.OpenAsync(cancellationToken);
-
-        await using (var command = Commands.GetErrorsXml(ApplicationName, errorIndex, pageSize))
+        var command = new NpgsqlCommand
         {
-            command.Connection = connection;
-
-            await using (var reader = await command.ExecuteReaderAsync(cancellationToken))
-            {
-                while (await reader.ReadAsync(cancellationToken))
-                {
-                    var id = reader.GetGuid(0);
-                    var xml = reader.GetString(1);
-                    var error = ErrorXml.DecodeString(xml);
-                    errorEntryList.Add(new ErrorLogEntry(this, id.ToString(), error));
-                }
-            }
-        }
-
-        await using (var command = Commands.GetErrorsXmlTotal(ApplicationName))
-        {
-            command.Connection = connection;
-            return Convert.ToInt32(await command.ExecuteScalarAsync(cancellationToken));
-        }
+            CommandText = "SELECT COUNT(*) FROM Elmah_Error WHERE Application = @Application"
+        };
+        command.Parameters.Add("@Application", NpgsqlDbType.Text, MaxAppNameLength).Value = appName;
+        return command;
     }
 
-    /// <summary>
-    ///     Creates the necessary tables and sequences used by this implementation
-    /// </summary>
-    private void CreateTableIfNotExists()
+    /// <inheritdoc />
+    protected override void CreateTableIfNotExists()
     {
         using var connection = new NpgsqlConnection(ConnectionString);
         connection.Open();
 
-        using var cmdCheck = Commands.CheckTable();
-        cmdCheck.Connection = connection;
-        // ReSharper disable once PossibleNullReferenceException
-        var exists = (bool)cmdCheck.ExecuteScalar();
-
-        if (exists) return;
-        using var cmdCreate = Commands.CreateTable();
-        cmdCreate.Connection = connection;
-        cmdCreate.ExecuteNonQuery();
-    }
-
-    private static class Commands
-    {
-        public static NpgsqlCommand CheckTable()
+        using var cmdCheck = new NpgsqlCommand
         {
-            var command = new NpgsqlCommand();
-            command.CommandText =
-                @"
+            Connection = connection,
+            CommandText = @"
 SELECT EXISTS (
    SELECT 1
-   FROM   information_schema.tables 
+   FROM   information_schema.tables
    WHERE  table_schema = 'public'
    AND    table_name = 'elmah_error'
    )
-";
-            return command;
-        }
+"
+        };
 
-        public static NpgsqlCommand CreateTable()
+        var exists = (bool)cmdCheck.ExecuteScalar()!;
+
+        if (!exists)
         {
-            var command = new NpgsqlCommand();
-            command.CommandText =
-                @"
+            using var cmdCreate = new NpgsqlCommand
+            {
+                Connection = connection,
+                CommandText = @"
 CREATE SEQUENCE ELMAH_Error_SEQUENCE;
 CREATE TABLE ELMAH_Error
 (
@@ -296,90 +186,9 @@ CREATE INDEX IX_ELMAH_Error_App_Time_Seq ON ELMAH_Error USING BTREE
     TimeUtc       DESC,
     Sequence      DESC
 );
-";
-
-            return command;
-        }
-
-        public static NpgsqlCommand LogError(
-            Guid id,
-            string appName,
-            string hostName,
-            string typeName,
-            string source,
-            string message,
-            string user,
-            int statusCode,
-            DateTime time,
-            string xml)
-        {
-            var command = new NpgsqlCommand();
-            command.CommandText =
-                @"
-/* elmah */
-INSERT INTO Elmah_Error (ErrorId, Application, Host, Type, Source, Message, ""User"", StatusCode, TimeUtc, AllXml)
-VALUES (@ErrorId, @Application, @Host, @Type, @Source, @Message, @User, @StatusCode, @TimeUtc, @AllXml)
-";
-            command.Parameters.Add(new NpgsqlParameter("ErrorId", id));
-            command.Parameters.Add(new NpgsqlParameter("Application", appName));
-            command.Parameters.Add(new NpgsqlParameter("Host", hostName));
-            command.Parameters.Add(new NpgsqlParameter("Type", typeName));
-            command.Parameters.Add(new NpgsqlParameter("Source", source));
-            command.Parameters.Add(new NpgsqlParameter("Message", message));
-            command.Parameters.Add(new NpgsqlParameter("User", user));
-            command.Parameters.Add(new NpgsqlParameter("StatusCode", statusCode));
-            command.Parameters.Add(new NpgsqlParameter("TimeUtc", time.ToUniversalTime()));
-            command.Parameters.Add(new NpgsqlParameter("AllXml", xml));
-
-            return command;
-        }
-
-        public static NpgsqlCommand GetErrorXml(string appName, Guid id)
-        {
-            var command = new NpgsqlCommand();
-
-            command.CommandText =
-                @"
-SELECT AllXml 
-FROM Elmah_Error 
-WHERE 
-    Application = @Application 
-    AND ErrorId = @ErrorId
-";
-
-            command.Parameters.Add(new NpgsqlParameter("Application", appName));
-            command.Parameters.Add(new NpgsqlParameter("ErrorId", id));
-
-            return command;
-        }
-
-        public static NpgsqlCommand GetErrorsXml(string appName, int errorIndex, int pageSize)
-        {
-            var command = new NpgsqlCommand();
-
-            command.CommandText =
-                @"
-SELECT ErrorId, AllXml FROM Elmah_Error
-WHERE
-    Application = @Application
-ORDER BY Sequence DESC
-OFFSET @offset
-LIMIT @limit
-";
-
-            command.Parameters.Add("@Application", NpgsqlDbType.Text, MaxAppNameLength).Value = appName;
-            command.Parameters.Add("@offset", NpgsqlDbType.Integer).Value = errorIndex;
-            command.Parameters.Add("@limit", NpgsqlDbType.Integer).Value = pageSize;
-
-            return command;
-        }
-
-        public static NpgsqlCommand GetErrorsXmlTotal(string appName)
-        {
-            var command = new NpgsqlCommand();
-            command.CommandText = "SELECT COUNT(*) FROM Elmah_Error WHERE Application = @Application";
-            command.Parameters.Add("@Application", NpgsqlDbType.Text, MaxAppNameLength).Value = appName;
-            return command;
+"
+            };
+            cmdCreate.ExecuteNonQuery();
         }
     }
 }
